fix(core): 修复全透明遮挡物致 _harden_occluder_alpha mode mismatch 崩溃
- 平铺胚衣(白W11/黑W11)的 occluder alpha 全为 0，原代码在 max==0 时
  保留 uint16 数组，Image.merge 与 L 模式通道不匹配报 ValueError 崩溃。
- 现无论 alpha 是否全 0 都 .astype(uint8) 归一成 L，崩溃消除。
- 副作用：单面款贴图现能正常生成平铺图(W白T/W黑T)，不再只出模特图。
- 顺带提交 cli.py 的 --occlusion-min-vis 参数(自适应遮挡保底，前次漏提交)。

+++ white_t_mockup/core.py
@@ -229,7 +229,10 @@ def _harden_occluder_alpha(occ: Image.Image) -> Image.Image:
     r, g, b, a = occ.split()
     a_arr = np.array(a, dtype=np.uint16)
     if a_arr.max() > 0:
-        a_arr = (a_arr * 255 // a_arr.max()).clip(0, 255).astype(np.uint8)
+        a_arr = (a_arr * 255 // a_arr.max()).clip(0, 255)
+    # 始终归一成 uint8(L 模式)：全透明遮挡物(alpha 最大值=0)若不转换会保留
+    # uint16(I;16)，与 r/g/b 的 L 模式不符导致 Image.merge 报 mode mismatch 崩溃。
+    a_arr = a_arr.astype(np.uint8)
     a = Image.fromarray(a_arr)
     return Image.merge("RGBA", (r, g, b, a))
